refactored project structure to better support the separation of client and server
