+++ src/client/mod.rs
@@ -1,0 +1,4 @@
+pub mod client_error;
+pub mod clientlog;
+pub mod race_event;
+pub mod logline_generator;
